+++ main-server/src/core/identity/add_accessor.js
@@ -29,7 +29,6 @@ import operationTypes from './operation_type';
 import * as tendermintNdid from '../../tendermint/ndid';
 import * as common from '../common';
 import * as cacheDb from '../../db/cache';
-import { getRequestMessageForAddingAccessor } from '../../utils/request_message';
 
 import CustomError from 'ndid-error/custom_error';
 import errorType from 'ndid-error/type';
@@ -71,6 +70,7 @@ export async function addAccessor(addAccessorParams) {
     identifier,
     accessor_type,
     accessor_public_key,
+    request_message,
   } = addAccessorParams;
 
   if (role === 'proxy') {
@@ -155,6 +155,11 @@ export async function addAccessor(addAccessorParams) {
     let request_id;
     if (mode === 3) {
       request_id = utils.createRequestId();
+      if (request_message == null) {
+        throw new CustomError({
+          errorType: errorType.REQUEST_MESSAGE_NEEDED,
+        });
+      }
     }
 
     await cacheDb.setIdentityRequestDataByReferenceId(node_id, reference_id, {
@@ -252,15 +257,7 @@ async function addAccessorInternalAsync(
           idp_id_list: [],
           callback_url: 'SYS_GEN_ADD_ACCESSOR',
           data_request_list: [],
-          request_message:
-            request_message != null
-              ? request_message
-              : getRequestMessageForAddingAccessor({
-                  namespace,
-                  identifier,
-                  reference_id,
-                  node_id: config.nodeId,
-                }),
+          request_message,
           min_ial: 1.1,
           min_aal: 1,
           min_idp,