+++ main-server/src/core/identity/revoke_accessor.js
@@ -27,7 +27,6 @@ import operationTypes from './operation_type';
 import * as common from '../common';
 import * as cacheDb from '../../db/cache';
 import * as tendermintNdid from '../../tendermint/ndid';
-import { getRequestMessageForRevokingAccessor } from '../../utils/request_message';
 
 import CustomError from 'ndid-error/custom_error';
 import errorType from 'ndid-error/type';
@@ -64,6 +63,7 @@ export async function revokeAccessor(revokeAccessorParams) {
     callback_url,
     namespace,
     identifier,
+    request_message,
   } = revokeAccessorParams;
 
   if (role === 'proxy') {
@@ -140,6 +140,11 @@ export async function revokeAccessor(revokeAccessorParams) {
     let request_id;
     if (mode === 3) {
       request_id = utils.createRequestId();
+      if (request_message == null) {
+        throw new CustomError({
+          errorType: errorType.REQUEST_MESSAGE_NEEDED,
+        });
+      }
     }
 
     await cacheDb.setIdentityRequestDataByReferenceId(node_id, reference_id, {
@@ -232,16 +237,7 @@ async function createRequestToRevokeAccessor(
           idp_id_list: [],
           callback_url: 'SYS_GEN_REVOKE_ACCESSOR',
           data_request_list: [],
-          request_message:
-            request_message != null
-              ? request_message
-              : getRequestMessageForRevokingAccessor({
-                  namespace,
-                  identifier,
-                  reference_id,
-                  node_id: config.nodeId,
-                  accessor_id,
-                }),
+          request_message,
           // WHAT SHOULD THESE BE? (IAL, AAL, MIN_IDP)
           min_ial: 1.1,
           min_aal: 1,