+++ main-server/src/core/identity/revoke_and_add_accessor.js
@@ -29,7 +29,6 @@ import operationTypes from './operation_type';
 import * as common from '../common';
 import * as cacheDb from '../../db/cache';
 import * as tendermintNdid from '../../tendermint/ndid';
-import { getRequestMessageForRevokingAndAddingAccessor } from '../../utils/request_message';
 
 import CustomError from 'ndid-error/custom_error';
 import errorType from 'ndid-error/type';
@@ -72,6 +71,7 @@ export async function revokeAndAddAccessor(revokeAndAddAccessorParams) {
     callback_url,
     namespace,
     identifier,
+    request_message,
   } = revokeAndAddAccessorParams;
 
   if (role === 'proxy') {
@@ -169,6 +169,11 @@ export async function revokeAndAddAccessor(revokeAndAddAccessorParams) {
     let request_id;
     if (mode === 3) {
       request_id = utils.createRequestId();
+      if (request_message == null) {
+        throw new CustomError({
+          errorType: errorType.REQUEST_MESSAGE_NEEDED,
+        });
+      }
     }
 
     await cacheDb.setIdentityRequestDataByReferenceId(node_id, reference_id, {
@@ -269,16 +274,7 @@ async function revokeAndAddAccessorInternalAsync(
           idp_id_list: [],
           callback_url: 'SYS_GEN_REVOKE_AND_ADD_ACCESSOR',
           data_request_list: [],
-          request_message:
-            request_message != null
-              ? request_message
-              : getRequestMessageForRevokingAndAddingAccessor({
-                  namespace,
-                  identifier,
-                  reference_id,
-                  node_id: config.nodeId,
-                  accessor_id,
-                }),
+          request_message,
           // WHAT SHOULD THESE BE? (IAL, AAL, MIN_IDP)
           min_ial: 1.1,
           min_aal: 1,