+++ main-server/src/core/identity/revoke_identity_association.js
@@ -27,7 +27,6 @@ import operationTypes from './operation_type';
 import * as tendermintNdid from '../../tendermint/ndid';
 import * as common from '../common';
 import * as cacheDb from '../../db/cache';
-import { getRequestMessageForRevokingAssociation } from '../../utils/request_message';
 
 import CustomError from 'ndid-error/custom_error';
 import errorType from 'ndid-error/type';
@@ -63,6 +62,7 @@ export async function revokeIdentityAssociation(
     callback_url,
     namespace,
     identifier,
+    request_message,
   } = revokeIdentityAssociationParams;
 
   if (role === 'proxy') {
@@ -118,6 +118,11 @@ export async function revokeIdentityAssociation(
     let request_id;
     if (identityOnNode.mode_list.includes(3)) {
       request_id = utils.createRequestId();
+      if (request_message == null) {
+        throw new CustomError({
+          errorType: errorType.REQUEST_MESSAGE_NEEDED,
+        });
+      }
     }
 
     await cacheDb.setIdentityRequestDataByReferenceId(node_id, reference_id, {
@@ -197,15 +202,7 @@ async function revokeAssociationInternalAsync(
           idp_id_list: [],
           callback_url: 'SYS_GEN_REVOKE_ASSOCIATION',
           data_request_list: [],
-          request_message:
-            request_message != null
-              ? request_message
-              : getRequestMessageForRevokingAssociation({
-                  namespace,
-                  identifier,
-                  reference_id,
-                  node_id: config.nodeId,
-                }),
+          request_message,
           min_ial: 1.1,
           min_aal: 1,
           min_idp,